+++ CHANGELOG.md
@@ -1,6 +1,21 @@
 Change-Log
 ===============
 
+### Release 1.0.2 ###
+> --.04.2017
+
+- Implementation of `BaseNavigationalTransition.configureTransitions(Activity)` has been split into
+  `configureIncomingTransitions(Activity)` and `configureOutgoingTransitions(Activity)`. The **outgoing**
+  configuration method is by default called whenever `BaseNavigationalTransition.start(Activity)` is 
+  invoked and the **incoming** configuration method should be called by the activity to which is the
+  calling activity transitioning from its `onCreate(Bundle)` method.
+- `BaseNavigationalTransition.configureTransitionsOverlapping(Activity)` has been deprecated as its
+  implementation has been moved into `configureIncomingTransitions(Activity)`.
+- `BaseNavigationalTransition` now also supports setting whether a transitioning shared elements should
+  use **overlay** or not via `sharedElementsUseOverlay(boolean)`.
+- Fixed **animation resource** for **enter back** animation of window transitions declared in 
+  `ExtranWindowTransitions` class
+
 ### Release 1.0.1 ###
 > 16.02.2017